+++ gateway-js/src/schema-helper/addExtensions.ts
@@ -1,13 +1,19 @@
 import { GraphQLSchema } from 'graphql';
+import { GraphQLSchemaExtensions } from 'graphql/type/schema';
+
 const { version } = require('../../package.json');
 
 export function addExtensions(schema: GraphQLSchema): GraphQLSchema {
+  const schemaExtension = schema.extensions as GraphQLSchemaExtensions ?? {};
+  const apolloExtension = schemaExtension?.apollo ?? {};
+  const gatewayExtension = apolloExtension?.gateway ?? {};
+
   schema.extensions = {
     ...schema.extensions,
     apollo: {
-      ...schema.extensions.apollo,
+      ...apolloExtension,
       gateway: {
-        ...schema.extensions.apollo?.gateway,
+        ...gatewayExtension,
         version,
       }
     },